requirements.txt
absl-py==2.1.0
accelerate==1.0.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==24.2.0
cachetools==5.5.0
calflops==0.3.2
certifi==2022.12.7
charset-normalizer==2.1.1
cmake==3.25.0
contourpy==1.1.1
cycler==0.12.1
einops==0.8.0
face-alignment==1.4.1
filelock==3.13.1
flatbuffers==24.3.25
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2024.10.0
gast==0.4.0
google-auth==2.36.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.67.1
h5py==3.11.0
hf-xet==1.1.5
huggingface-hub==0.33.4
idna==3.4
imageio==2.35.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
Jinja2==3.1.3
keras==2.13.1
kiwisolver==1.4.7
lazy_loader==0.4
libclang==18.1.1
lit==15.0.7
llvmlite==0.41.1
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
multidict==6.1.0
networkx==3.0
numba==0.58.1
numpy==1.24.1
oauthlib==3.2.2
opencv-python==4.10.0.84
opt_einsum==3.4.0
packaging==24.1
pandas==2.0.3
pillow==10.2.0
propcache==0.2.0
protobuf==4.25.5
psutil==6.1.0
ptflops==0.7.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytorch-msssim==1.0.0
pytz==2024.2
PyWavelets==1.4.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.28.1
requests-oauthlib==2.0.0
rsa==4.9
safetensors==0.5.3
scikit-image==0.21.0
scipy==1.10.1
six==1.16.0
sympy==1.13.1
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorboard-logger==0.1.0
tensorflow==2.13.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
tifffile==2023.7.10
tokenizers==0.20.3
torch==2.0.1
torch-geometric==2.6.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.66.6
transformers==4.46.3
triton==2.0.0
typing_extensions==4.5.0
tzdata==2024.2
urllib3==1.26.13
Werkzeug==3.0.6
wrapt==1.16.0
yarl==1.15.2
zipp==3.20.2
